requirements.txt
py-cord~=2.6.1
python-dotenv~=1.1.0
qrcode~=8.0
peewee~=3.17.9
aiohttp~=3.11.14